+++ R/map_layer_pointcloud.R
@@ -61,10 +61,10 @@ mapdeckPointcloudDependency <- function() {
 #' )
 #'
 #' ## as an sf object wtih a Z attribute
-#' library(sf)
-#' sf <- sf::st_as_sf( df , coords = c("lon","lat","z"))
+#' library(sfheaders)
+#' sf <- sfheaders::sf_point( df, x = "lon", y = "lat", z = "z" )
 #'
-#' mapdeck(token = key, style = mapdeck_style("dark")) %>%
+#' mapdeck(style = mapdeck_style("dark")) %>%
 #' add_pointcloud(
 #'   data = sf
 #'   , layer_id = 'point'
@@ -122,7 +122,7 @@ add_pointcloud <- function(
 	l <- resolve_palette( l, palette )
 	l <- resolve_legend( l, legend )
 	l <- resolve_legend_options( l, legend_options )
-	l <- resolve_elevation_data( data, l, elevation, c("POINT","MULTIPOINT") )
+	l <- resolve_elevation_data( data, l, elevation, c("POINT") )
 
 	bbox <- init_bbox()
 	update_view <- force( update_view )
@@ -164,7 +164,7 @@ add_pointcloud <- function(
 			l[["elevation"]] <- 0
 		}
 
-		print( head( data )  )
+		#print( head( data )  )
 		geometry_column <- list( geometry = c("lon", "lat","elevation") )
 		shape <- rcpp_point_df_columnar( data, l, geometry_column, digits, "pointcloud" )