+++ inst/htmlwidgets/lib/arc/arc.js
@@ -3,7 +3,7 @@ function add_arc_geo( map_id, map_type, arc_data, layer_id, auto_highlight, high
   var extensions = [];
 
   if ( brush_radius > 0 ) {
-  	extensions.push( new BrushingExtension() );
+  	extensions.push( new deck.BrushingExtension() );
   }
 
   const arcLayer = new deck.ArcLayer({
@@ -34,7 +34,7 @@ function add_arc_geo( map_id, map_type, arc_data, layer_id, auto_highlight, high
 	}
 
 	if (legend !== false) {
-	  md_add_legend( map_id, map_type, layer_id, legend );
+	  md_add_legend( map_id, map_type, layer_id, legend, "hex" );
 	}
 	md_layer_view( map_id, map_type, layer_id, focus_layer, bbox, update_view );
 }
@@ -45,7 +45,7 @@ function add_arc_polyline( map_id, map_type, arc_data, layer_id, auto_highlight,
   var extensions = [];
 
   if ( brush_radius > 0 ) {
-  	extensions.push( new BrushingExtension() );
+  	extensions.push( new deck.BrushingExtension() );
   }
 
   const arcLayer = new deck.ArcLayer({
@@ -76,7 +76,7 @@ function add_arc_polyline( map_id, map_type, arc_data, layer_id, auto_highlight,
 	}
 
 	if (legend !== false) {
-	  md_add_legend( map_id, map_type, layer_id, legend );
+	  md_add_legend( map_id, map_type, layer_id, legend, "hex" );
 	}
 	md_layer_view( map_id, map_type, layer_id, focus_layer, bbox, update_view );
 }