+++ inst/htmlwidgets/lib/polygon/polygon.js
@@ -1,5 +1,11 @@
 
-function add_polygon_geo( map_id, map_type, polygon_data, layer_id, light_settings, auto_highlight, highlight_colour, legend, bbox, update_view, focus_layer, js_transition, is_extruded, elevation_scale ) {
+function add_polygon_geo( map_id, map_type, polygon_data, layer_id, light_settings, auto_highlight, highlight_colour, legend, bbox, update_view, focus_layer, js_transition, is_extruded, elevation_scale, brush_radius ) {
+
+  var extensions = [];
+
+  if ( brush_radius > 0 ) {
+  	extensions.push( new deck.BrushingExtension() );
+  }
 
   const polygonLayer = new deck.PolygonLayer({
   	map_id: map_id,
@@ -22,7 +28,9 @@ function add_polygon_geo( map_id, map_type, polygon_data, layer_id, light_settin
     highlightColor: md_hexToRGBA( highlight_colour ),
     onHover: md_update_tooltip,
     onClick: info => md_layer_click( map_id, "polygon", info ),
-    transitions: js_transition || {}
+    transitions: js_transition || {},
+    brushingRadius: brush_radius,
+    extensions: extensions
   });
 
   if( map_type == "google_map") {
@@ -33,13 +41,19 @@ function add_polygon_geo( map_id, map_type, polygon_data, layer_id, light_settin
   }
 
 	if (legend !== false) {
-	  md_add_legend(map_id, map_type, layer_id, legend);
+	  md_add_legend(map_id, map_type, layer_id, legend, "hex" );
 	}
 	md_layer_view( map_id, map_type, layer_id, focus_layer, bbox, update_view );
 }
 
 
-function add_polygon_polyline( map_id, map_type, polygon_data, layer_id, light_settings, auto_highlight, highlight_colour, legend, bbox, update_view, focus_layer, js_transition, is_extruded, elevation_scale ) {
+function add_polygon_polyline( map_id, map_type, polygon_data, layer_id, light_settings, auto_highlight, highlight_colour, legend, bbox, update_view, focus_layer, js_transition, is_extruded, elevation_scale, brush_radius ) {
+
+  var extensions = [];
+
+  if ( brush_radius > 0 ) {
+  	extensions.push( new deck.BrushingExtension() );
+  }
 
   const polygonLayer = new deck.PolygonLayer({
     map_id: map_id,
@@ -62,7 +76,9 @@ function add_polygon_polyline( map_id, map_type, polygon_data, layer_id, light_s
     highlightColor: md_hexToRGBA( highlight_colour ),
     onHover: md_update_tooltip,
     onClick: info => md_layer_click( map_id, "polygon", info ),
-    transitions: js_transition || {}
+    transitions: js_transition || {},
+    brushingRadius: brush_radius,
+    extensions: extensions
   });
 
   if( map_type == "google_map") {
@@ -72,7 +88,7 @@ function add_polygon_polyline( map_id, map_type, polygon_data, layer_id, light_s
   }
 
 	if (legend !== false) {
-	  md_add_legend(map_id, map_type, layer_id, legend);
+	  md_add_legend(map_id, map_type, layer_id, legend, "hex" );
 	}
 	md_layer_view( map_id, map_type, layer_id, focus_layer, bbox, update_view );
 }